Restructured GUI code.
DukeDialogbox, UserDialogbox and MainWindow are now in package duke.gui.
