+++ data/savedTasks.txt
@@ -2,5 +2,4 @@ D | 1 | project work | 26/02/1997 18:00
 T | 1 | ProjectWork
 T | 1 | Compelte CCA Form
 E | 1 | Orientation | 31/08/2019 09:00 | 31/08/2019 09:00
-T | 1 | Even More Stuff
-T | 0 | add more stuff
+T | 1 | Complete CCA Form
